container-app/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai>=1.0.0
azure-identity==1.19.0
httpx==0.25.2
pydantic==2.5.0
